requirements.txt
wasmer==1.0.0
wasmer_compiler_cranelift
paho-mqtt
pyserial